storage: move storage_state module to protocol::server::worker
I think this module movement was just missed in #13626. It seems to
logically belong in the server module introduced in #13626.
