+++ src/storage/src/sink/kafka.rs
@@ -59,8 +59,8 @@ use mz_timely_util::operators_async_ext::OperatorBuilderExt;
 
 use super::KafkaBaseMetrics;
 use crate::controller::CollectionMetadata;
+use crate::protocol::server::worker::StorageState;
 use crate::render::sinks::SinkRender;
-use crate::storage_state::StorageState;
 use crate::types::connections::{ConnectionContext, PopulateClientConfig};
 use crate::types::errors::DataflowError;
 use crate::types::sinks::{
